+++ Common/src/main/java/com/unrealdinnerbone/trenzalore/Trenzalore.java
@@ -3,8 +3,10 @@ package com.unrealdinnerbone.trenzalore;
 import com.unrealdinnerbone.trenzalore.api.platform.Services;
 import com.unrealdinnerbone.trenzalore.api.platform.services.IRegistry;
 import com.mojang.logging.LogUtils;
+import com.unrealdinnerbone.trenzalore.api.registry.RegistryObjects;
 import org.slf4j.Logger;
 
+import java.util.List;
 import java.util.ServiceLoader;
 
 public class Trenzalore {
@@ -14,7 +16,10 @@ public class Trenzalore {
 
     public static void init() {
         ServiceLoader.load(IRegistry.class)
-                        .forEach(iRegistry -> iRegistry.getRegistryObjects()
-                                .forEach(registryObject -> Services.INTERNAL.registryRegistryObjects(iRegistry.getModID(), registryObject)));
+                        .forEach(iRegistry -> {
+                            List<RegistryObjects<?>> registryObjects = iRegistry.getRegistryObjects();
+                            registryObjects
+                                    .forEach(registryObject -> Services.PLATFORM.registryRegistryObjects(iRegistry.getModID(), registryObject));
+                        });
     }
 }