+++ Common/src/main/java/com/unrealdinnerbone/trenzalore/api/platform/services/IRegistry.java
@@ -8,4 +8,6 @@ public interface IRegistry {
     List<RegistryObjects<?>> getRegistryObjects();
 
     String getModID();
+
+
 }