+++ Common/src/main/java/com/unrealdinnerbone/trenzalore/api/registry/RegistryFactory.java
@@ -1,7 +1,11 @@
 package com.unrealdinnerbone.trenzalore.api.registry;
 
+import com.unrealdinnerbone.trenzalore.api.platform.Services;
+import net.minecraft.advancements.CriterionTrigger;
 import net.minecraft.core.Registry;
 import net.minecraft.resources.ResourceKey;
+import net.minecraft.world.item.CreativeModeTab;
+import net.minecraft.world.item.Item;
 import org.jetbrains.annotations.ApiStatus;
 
 import java.util.ArrayList;
@@ -11,24 +15,21 @@ import java.util.Map;
 
 public class RegistryFactory {
 
-    @ApiStatus.Internal
-    private static final Map<String, List<RegistryObjects<?>>> registryMap = new HashMap<>();
-
+    @Deprecated(forRemoval = true)
+    @ApiStatus.ScheduledForRemoval(inVersion = "3.0.0")
     public static <T> RegistryObjects<T> create(String modid, ResourceKey<Registry<T>> registry) {
-        if(!registryMap.containsKey(modid)) {
-            registryMap.put(modid, new ArrayList<>());
-        }
-        if(registryMap.get(modid).stream().anyMatch(registryObjects -> registryObjects.registryKey().equals(registry))) {
-            throw new RuntimeException("Registry already exists");
-        }
-        RegistryObjects<T> registryObjects = new RegistryObjects<>(registry, new ArrayList<>());
-        registryMap.get(modid).add(registryObjects);
-        return registryObjects;
-
+        return new RegistryObjects<>(registry, new ArrayList<>());
     }
 
-    @ApiStatus.Internal
-    public static Map<String, List<RegistryObjects<?>>> getRegistryMap() {
-        return registryMap;
+    public static <T> RegistryObjects<T> create(ResourceKey<Registry<T>> registry) {
+        return new RegistryObjects<>(registry, new ArrayList<>());
+    }
+
+    public static <T extends CriterionTrigger<?>> T registerCriterion(T criterion) {
+        return Services.PLATFORM.registerCriterion(criterion);
+    }
+
+    public static void registerCreativeTabItems(ResourceKey<CreativeModeTab> tabResourceKey, List<Item> item) {
+        Services.PLATFORM.addItemToCreativeTab(tabResourceKey, item);
     }
 }