+++ Fabric/src/main/java/com/unrealdinnerbone/trenzalore/platform/FabricPlatformHelper.java
@@ -1,9 +1,27 @@
 package com.unrealdinnerbone.trenzalore.platform;
 
 import com.unrealdinnerbone.trenzalore.api.platform.services.IPlatformHelper;
+import com.unrealdinnerbone.trenzalore.api.registry.RegistryEntry;
+import com.unrealdinnerbone.trenzalore.api.registry.RegistryObjects;
+import net.fabricmc.fabric.api.dimension.v1.FabricDimensions;
+import net.fabricmc.fabric.api.itemgroup.v1.ItemGroupEvents;
 import net.fabricmc.loader.api.FabricLoader;
+import net.minecraft.advancements.CriteriaTriggers;
+import net.minecraft.advancements.CriterionTrigger;
+import net.minecraft.core.Registry;
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.item.CreativeModeTab;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.level.portal.PortalInfo;
+import org.jetbrains.annotations.ApiStatus;
+import org.jetbrains.annotations.Nullable;
 
 import java.nio.file.Path;
+import java.util.List;
 
 public class FabricPlatformHelper implements IPlatformHelper {
     @Override
@@ -14,6 +32,32 @@ public class FabricPlatformHelper implements IPlatformHelper {
     @Override
     public Path getConfigPath() {
         return FabricLoader.getInstance().getConfigDir();
+    }
+
+    @Override
+    public <T extends CriterionTrigger<?>> T registerCriterion(T criterion) {
+        return CriteriaTriggers.register(criterion);
+    }
+
+    @Override
+    @ApiStatus.Internal
+    public void addItemToCreativeTab(ResourceKey<CreativeModeTab> tabResourceKey, List<Item> items) {
+        ItemGroupEvents.modifyEntriesEvent(tabResourceKey).register(entries -> items.forEach(entries::accept));
+    }
+
+    @Override
+    public <T> void registryRegistryObjects(String modId, RegistryObjects<T> registryObjects) {
+        ResourceKey<Registry<T>> registryKey = registryObjects.registryKey();
+        Registry<T> registry = (Registry<T>) BuiltInRegistries.REGISTRY.get(registryKey.location());
+        for (RegistryEntry<T> object : registryObjects.objects()) {
+            ResourceLocation id = new ResourceLocation(modId, object.name());
+            Registry.register(registry, id, object.get());
+        }
+    }
+
+    @Override
+    public @Nullable <T extends Entity> Entity teleport(T entity, ServerLevel level, PortalInfo portalInfo) {
+        return FabricDimensions.teleport(entity, level, portalInfo);
     }
 
 }