+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### 2.1.0
+
+- Add Creative Tab Helper
+- Cleanup Services into one
+- Added Helper to register Criterion
+- made it, so you don't need a mod id to create a RegistryObjects
+
 ### 2.0.4
 
 - Fixed Jar File Name